PrintMemoryWorking

+++ Userland/Include/String.h
@@ -1,11 +1,14 @@
 #ifndef STRINGH
 #define STRINGH
 
+#include <stdint.h>
+
+
 int strlen(char * str);
 
-void IntToString(char * buff, int buffSize, int num);
+void IntToString(char * buff, int buffSize, uint64_t num);
 
-void HexToString(char * buff, int buffSize, int num);
+void HexToString(char * buff, int buffSize, uint64_t num);
